fix(fixtures): Move queries to a separate module
