+++ fixtures/queries/__init__.py
@@ -1,0 +1,11 @@
+def escape_string(s):
+    if s is None:
+        return s
+
+    replacements = {
+        "'": "''",
+        '"': '""',
+    }
+    for old, new in replacements.items():
+        s = s.replace(old, new)
+    return s
