+++ fixtures/queries/questions.py
@@ -1,35 +1,3 @@
-def escape_string(s):
-    if s is None:
-        return s
-
-    replacements = {
-        "'": "''",
-        '"': '""',
-    }
-    for old, new in replacements.items():
-        s = s.replace(old, new)
-    return s
-
-
-def get_subject_id(subject, conn):
-    with conn.cursor() as cur:
-        cur.execute("""SELECT id FROM subjects WHERE name = '{}'""".format(subject))
-        subject_id = cur.fetchone()
-
-    return subject_id[0] if subject_id is not None else None
-
-
-def get_or_insert_subject(subject, conn):
-    with conn.cursor() as cur:
-        if get_subject_id(subject, conn) is None:
-            cur.execute(
-                """INSERT INTO subjects (name, inserted_at, updated_at) VALUES ('{}', now(), now());""".format(subject)
-            )
-        conn.commit()
-
-    return get_subject_id(subject, conn)
-
-
 def get_question_id(subject_id, content, image, conn):
     with conn.cursor() as cur:
         query = """SELECT id FROM questions WHERE content = '{}' AND image_name = {} AND subject_id = '{}'""".format(
@@ -51,10 +19,11 @@ def get_or_insert_question(question, conn):
 
     with conn.cursor() as cur:
         if get_question_id(subject_id, content, image, conn) is None:
-            print(content)
             query = """INSERT INTO questions (subject_id, content, image_name, inserted_at, updated_at) VALUES ({}, '{}', {}, now(), now());""".format(
                 subject_id, content, image)
             cur.execute(query)
-        conn.commit()
-        cur.close()
+            conn.commit()
+        else:
+            print(question, content, image, subject_id)
+
     return get_question_id(subject_id, content, image, conn)
